JSON ARRAPPEN ARRLEN ARRPOP refractor
